+++ Windows Tkinter App/main.py
@@ -1,4 +1,4 @@
-import menu
+from rfid_app import MyApp
 
 if __name__ == "__main__":
-    menu.run()
+    MyApp().run()
